+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [v1.5.8] - 2018-03-??
 
-- TODO: write change before preparing next tag
+- Change all `getRef` with `innerRef`.
 
 ## [v1.5.7] - 2018-03-27